New Rovereto policy and updated max duplicate per TType value.

+++ src/main/java/eu/trentorise/smartcampus/mobility/controller/extensions/PlanRequest.java
@@ -15,7 +15,7 @@ public class PlanRequest {
 	private RType routeType;
 	private String request;
 	private String plan;
-	private Integer value;
+	private Double value;
 	private List<Itinerary> itinerary;
 	private SingleJourney originalRequest;
 	private int itineraryNumber;
@@ -52,10 +52,10 @@ public class PlanRequest {
 		this.plan = plan;
 	}
 
-	public Integer getValue() {
+	public Double getValue() {
 		return value;
 	}
-	public void setValue(Integer value) {
+	public void setValue(Double value) {
 		this.value = value;
 	}
